getLogPos() should return an int

+++ trunk/h2/src/main/org/h2/store/PageLog.java
@@ -687,7 +687,7 @@ public class PageLog {
         return firstSectionId;
     }
 
-    long getLogPos() {
+    int getLogPos() {
         return logPos;
     }
